Make SVN::Look::proplist support SVN 1.8
SVN 1.8 changed the output format of the proplist command. So, in
order to make the code more stable we try to use the --xml
option. However, this option was implemented on svn 1.6.0 only. Since
we still want to support older Subversions we have to check if we can
use that option first.

Thanks to George Hartzell to pointing this out to me at
https://github.com/gnustavo/SVN-Look/pull/1.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>